@@ -16,10 +16,10 @@ Gem::Specification.new do |s|
 
   s.files = Dir['{app,config,db,lib}/**/*', 'MIT-LICENSE', 'Rakefile', 'README.rdoc']
 
-  s.add_dependency 'rails', '>= 4.2'
-  s.add_dependency 'activejob', '>= 4.2'
+  s.add_dependency 'rails', '>= 5.0.2'
+  s.add_dependency 'activejob', '>= 5.0.2'
   s.add_dependency 'responders', '~> 2.0'
-  s.add_dependency 'mercurius', '>= 0.2.0'
+  s.add_dependency 'mercurius', '>= 0.2.1'
 
   s.add_development_dependency 'sqlite3'
   s.add_development_dependency 'rspec-rails'
